Cleanup how various elements are serialized when logging
Added a new interface FeideConnect\Utils\Loggable that defines a method to get a log representation of a object. The Logger class looks for instances that implements this interface and converts the data using that method. This means we don't have to worry about how to convert a object to logable representation everywhere we want to log that object