Reworked passive authentication
Now we iterate over over all authsources and check if they have an
active session, if se we use that. If no sessions are found we send a
passive authentication request to the default auth source which would
be feide. In order to be allowed to do a passive authentication the
client must have requireInteraction set to false in authoptions