Account: Rework validateAuthProvider() logic.
This patch changes the account access check logic in validateAuthProvider().
The new method simply finds all possibilities from the current acount, and
intersect them with the ones allowed by the service. If the intersection it
means that the current account is allowed for this service, and the login
is allowed